move cdm/src/timing to cdm-test/src/timing
