+++ cdm-test/src/timing/java/timing/Average.java
