+++ cdm-test/src/timing/java/ucar/nc2/util/xml/TestMetarEncoding.java
